Emulator working on AVR

+++ platform/avr/spi/spi.c
@@ -38,7 +38,7 @@ u8 spi_xchg(u8 byte)
 
 Status spi_xchg_try(u8 in, u8 *out)
 {
-	u32 cnt = 0;
+	u16 cnt = 0;
 	SPDR = in;
 	while(!(SPSR & (1 << SPIF)))
 	{
